pyblocks: Improve code structure
Claim the board lock and check if coordinates in one place in
update_board_list(), instead of in multiple places where
update_board_list() is used.

Claim the board lock when updating the board for filled
lines to prevent conflicts between the main thread and
input handler thread.

Create a wrapper function for the input_handler_callback()
to separate checking for a 'QUIT' event from handling quad
translations.

Signed-off-by: Zayd Osman <zaydosman@gmail.com>